feat(intelli-espower-loader): add default patterns
